Added a bouncy tile and a stretch animation. Fixed collision bugs.

+++ src/com/rb2750/lwjgl/world/WorldSettings.java
@@ -6,11 +6,11 @@ import lombok.Setter;
 public class WorldSettings {
     @Getter
     @Setter
-    private double gravity = 1.2f;
+    private float gravity = 1.2f;
     @Getter
     @Setter
-    private double frictionGround = 0.6f;
+    private float frictionGround = 0.6f;
     @Getter
     @Setter
-    private double frictionAir = 0.1f;
+    private float frictionAir = 0.1f;
 }
